requirements/base.txt
pytz==2022.1  # https://github.com/stub42/pytz
python-slugify==6.1.2  # https://github.com/un33k/python-slugify
Pillow==9.1.1  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
redis==4.3.3  # https://github.com/redis/redis-py
hiredis==2.0.0  # https://github.com/redis/hiredis-py
uvicorn[standard]==0.17.6  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django==3.2.13  # pyup: < 4.0  # https://www.djangoproject.com/
django-environ==0.8.1  # https://github.com/joke2k/django-environ
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.50.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.2.0  # https://github.com/jazzband/django-redis
django-compressor==4.0  # https://github.com/django-compressor/django-compressor

# Django REST Framework
djangorestframework==3.13.1  # https://github.com/encode/django-rest-framework
django-cors-headers==3.11.0  # https://github.com/adamchainz/django-cors-headers


# DRF-spectacular for api documentation
drf-spectacular==0.22.1  # https://github.com/tfranzel/drf-spectacular


# JWT
djangorestframework-simplejwt==5.1.0  # https://django-rest-framework-simplejwt.readthedocs.io/en/latest/getting_started.html


# Django Reversion
django-reversion==5.0.0  # https://github.com/etianen/django-reversion


# Control Live field (https://github.com/hearsaycorp/django-livefield)
django-livefield==3.3.0


# https://github.com/KristianOellegaard/django-health-check
django-health-check==3.16.5


# Django Filer
# https://django-filer.readthedocs.io/en/latest/index.html
django-filer==2.0.2


# psutil
# https://github.com/giampaolo/psutil
psutil==5.9.0
